implement background scrolling
Squashed commit of the following:

commit cc208ee3fc5535e4808c0cf0eee34a3ca825582c
Author: bmoxb <maxoblack@yahoo.com>
Date:   Sun Jul 30 20:53:51 2023 +0200

    fix scrolling clipping at edge of screen

commit 609bd89a9eec9a2bfae855d95a6ea966266966e4
Author: bmoxb <maxoblack@yahoo.com>
Date:   Sun Jul 30 12:15:02 2023 +0200

    add vertical scrolling

commit 5cc11f3b615ffca2e8eefe52755114db748521c6
Author: bmoxb <maxoblack@yahoo.com>
Date:   Sat Jul 29 23:20:14 2023 +0200

    implement horizontal scroll

+++ core/src/gpu/vram.rs
@@ -67,7 +67,7 @@ impl VideoRam {
     }
 
     fn read_tile_index(&self, offset: u16, x: u8, y: u8) -> u8 {
-        debug_assert!((x as usize) < TILE_MAP_WIDTH && (y as usize) < TILE_MAP_WIDTH);
+        debug_assert!(x < TILE_MAP_WIDTH as u8 && y < TILE_MAP_WIDTH as u8);
         self.read8(offset + (y as u16 * TILE_MAP_WIDTH as u16) + x as u16)
     }
 }
